Fix and improvements
"Fase4":

- Fixed an issue where the "Max range to detect scene change before the initial timestamp (ms)" and "Max range to detect scene change after the initial timestamp (ms)" values in the GUI were not applying correctly, remaining fixed at the hardcoded base value (fallback).

- Changed the hardcoded values to make them slightly more aggressive. This function looks for lines where the initial timestamp is too close to a speech peak (which could cause them to appear delayed visually on video) and inserts a lead-in to compensate for the delay.
Consequently, the "leftovers" given by "Fase2" will be corrected where possible by "Fase4".

(Now that the values are applied correctly, you need to take into account the lead-in set in the GUI.
For example: Lead-in 180 (Fase 2), Max range 250 (Fase 4, third setting). That is, 180±250, so a maximum of 430ms is possible if a scene change is detected.)

"Fase3":

The timestamps of the lines are now more accurate to the frame (23,976).

It was previously on the correct frame, but it was a few ms early or late (although it looked correct on the video player).

For 24,000 fps or other frame rates, it should now work, but you'll likely need to do a global shift for all lines to accommodate pyscenedetect delays or advances. (Not fully tested.)